requirements.txt
antiorm==1.2.1
anyascii==0.1.7
asgiref==3.3.1
autopep8==1.5.4
beautifulsoup4==4.8.2
certifi==2020.12.5
chardet==4.0.0
db==0.1.1
db-sqlite3==0.0.1
Django==3.0
django-admin-thumbnails==0.2.5
django-appconf==1.0.4
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-debug-toolbar==3.2
django-filter==2.4.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-modelcluster==5.1
django-nested-admin==3.3.3
django-taggit==1.3.0
django-treebeard==4.4
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
html5lib==1.1
idna==2.10
jdcal==1.4.1
l18n==2020.6.1
openpyxl==3.0.6
pilkit==2.0
Pillow==8.1.0
psycopg2==2.8.6
pycodestyle==2.6.0
pycryptodome==3.10.1
python-monkey-business==1.0.0
pytz==2020.5
razorpay==1.2.0
requests==2.25.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
urllib3==1.26.3
wagtail==2.12.2
webencodings==0.5.1
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
